Extracted MyBus UI models classes to files
